Change the SingleGeoQuery
The code of the single geo query is moved in a single function inside GeoFirestore class like the method for getting/setting/removing a geo location.
The method, like the others, has an extension function that accept a lambda for Kotlin users.
