+++ geofirestore/src/main/java/org/imperiumlabs/geofirestore/listeners/GeoQueryDataEventListener.kt
@@ -1,4 +1,4 @@
-package org.imperiumlabs.geofirestore.callbacks
+package org.imperiumlabs.geofirestore.listeners
 
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.GeoPoint
